add webserver.
This adds a yesod webserver to serve the results from the
sqlite database.

Currently the links to files are still broken but I need to deploy
to fix that correctly.
(because then it uses the nixified sqlite database)

Something to be said to not put this in the main program
because it incurs a lot of dependencies.
But I thought to get something out faster it's better
to create a large monolith,
deploy it and figure out how to clean it once (or if)
people actually use it.

![image](https://user-images.githubusercontent.com/3874017/166161805-f58309b5-cf7c-4fda-8f65-2a05fbe69896.png)

![image](https://user-images.githubusercontent.com/3874017/166162029-d990602e-de9f-41ce-865c-295810c66bee.png)